[🆕] NT-253 Stripe SDK 8.3.0 -> 11.1.3 (#618)
Updated the Stripe SDK to 11.1.3
Had to update the gradle wrapper to 5.1.1 and gradle build tools to 3.4.2
Had to update Robolectric to 4.3 and remove unused SDKs
Hade to update Apollo library to 1.0.2
Apollo now explicitly wants an Integer for the result of the ClearUnseenActivityMutation.
Updated usage in MockStripe, NewCardFragment, NewCardFragmentViewModel, NewCardFragmentViewModelTest

+++ app/src/main/java/com/kickstarter/mock/services/MockApolloClient.kt
@@ -26,8 +26,8 @@ open class MockApolloClient : ApolloClientType {
         return Observable.just(true)
     }
 
-    override fun clearUnseenActivity(): Observable<Long> {
-        return Observable.just(0L)
+    override fun clearUnseenActivity(): Observable<Int> {
+        return Observable.just(0)
     }
 
     override fun createPassword(password: String, confirmPassword: String): Observable<CreatePasswordMutation.Data> {
